Generative-AI/requirements.txt
streamlit ## used for frontend
google-generativeai ## used for Google Generative AI
python-dotenv ## used for managing environment variables

